refactor: 字幕翻译网关按引擎拆分为云端与本地两个实现
业务层不再拼提示词，只向网关传语义输入（当前组、组前后句、
模式、风格）；提示词模板与结构化输出拼装下放到基础设施，
本机与在线两条链路彻底分离，为本地引擎自有 fallback 留出落点。

- 新增 OpenAiSubtitleBatchTranslator / LocalSubtitleBatchTranslator
  两个网关契约与实现；云端保持 40s 超时 + gpt 限流，本地保持经
  LocalAiRuntime 的串行推理，两侧超时/重试/容错策略独立演进
- 提示词模板、translation 字段说明与结果 schema 收敛到
  infrastructure/translate/subtitleBatchPrompt.ts，两引擎当前共用，
  之后可按各自模型特性分叉
- 引擎路由仍由业务层按已解析设置决定（localModelId 判空），
  腾讯批量翻译路径不变；validateGatewayItems 暂留业务层
- 删除统一入口 SubtitleTranslationGateway；openaiSubtitlePrompts.ts
  仅保留风格默认值、设置键与风格签名
