+++ src/backend/ioc/types.ts
@@ -59,7 +59,8 @@ const TYPES = {
     YouDaoClientProvider: Symbol('YouDaoClientProvider'),
     TencentClientProvider: Symbol('TencentClientProvider'),
     AiProviderService: Symbol('AiProviderService'),
-    SubtitleTranslationGateway: Symbol('SubtitleTranslationGateway'),
+    OpenAiSubtitleBatchTranslator: Symbol('OpenAiSubtitleBatchTranslator'),
+    LocalSubtitleBatchTranslator: Symbol('LocalSubtitleBatchTranslator'),
     ModelRoutingService: Symbol('ModelRoutingService'),
     ConfigStoreFactory: Symbol('ConfigStoreFactory'),
     SettingsStore: Symbol('SettingsStore'),
